requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
llama-index-core>=0.12.0,<0.13.0
llama-index-llms-deepseek==0.1.1
